Alpha implementation of ML media lookup and API.
Quite a bit of work went into the API, which allows for basic searching
and generation of assets.

Also included are tests, one of which includes CC-BY licensed music. The
attribution is in "media-info.txt" in the test-files folder. This is not
an endorsement of the music, I just needed a test file that I could
include in the repo.